refactor: extract xtdb schema
