+++ polars_hist_db/overrides/arrow_xtdb.py
@@ -8,10 +8,8 @@ import polars as pl
 import pyarrow as pa
 from sqlalchemy import text
 
-from polars_hist_db.backends.xtdb import (
-    _execute_xtdb_transaction,
-    _xtdb_table_exists,
-)
+from polars_hist_db.backends.xtdb import _execute_xtdb_transaction
+from polars_hist_db.backends.xtdb_schema import _xtdb_table_exists
 from polars_hist_db.backends.xtdb_transport import _xtdb_column_identifier
 
 from .arrow import (
